requirements.txt
flasgger==0.9.7.1
pip==19.3.1
Flask==3.0.3
Flask-RESTful==0.3.10
stripe==11.2.0
SQLAlchemy==2.0.32
requests==2.31.0
Werkzeug==3.0.4
marshmallow==3.22.0
Flask-SQLAlchemy==3.0.2
gunicorn==20.1.0
PyJWT==2.9.0
python-dotenv==1.0.1 
pillow==11.0.0 
PyAutoGUI==0.9.54
pandas==2.1.1
selenium==4.14.0
wheel==0.37.1
demjson3==3.0.6 
setuptools==58.0.0
anyjson==0.3.3 
celery==5.4.0 